% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contact_all}
\alias{contact_all}
\title{contact_all.rda}
\format{
A list with 177 elements representing unique countries
}
\usage{
data(contact_all)
}
\description{
A data set
}
\keyword{datasets}
