% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertActivitySIA.R
\name{ConvertActivitySIA}
\alias{ConvertActivitySIA}
\title{Covert SIA activities into 4 groupings and exclude (or not) based on logicals}
\usage{
ConvertActivitySIA(df_sia, PIRI, OR, OTHER)
}
\arguments{
\item{df_sia}{- dataframe}

\item{PIRI}{- logical - to include SIAs Activities focused on RI or not}

\item{OR}{- logical - to include outbreak response SIAs Activities or not}

\item{OTHER}{- logical - to include any other SIAs Activities or not}
}
\value{
list of new dataframe AND vector of excluded SIAs
}
\description{
Covert SIA activities into 4 groupings and exclude (or not) based on logicals
}
