% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertAgeSIA.R
\name{convertAgeSIA}
\alias{convertAgeSIA}
\title{Function to convert WHOs SIA age range, as text, into something usable}
\usage{
convertAgeSIA(sia.age.range, check = FALSE)
}
\arguments{
\item{sia.age.range}{input as AgeGroup from the WHO SIA Data}

\item{check}{boolean of whether to print diagnostics of the functions success, default is FALSE}
}
\value{
Returns age range in months
}
\description{
Function to convert WHOs SIA age range, as text, into something usable
}
