% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertAgeSIA_v2.R
\name{ConvertAgeSIA_v2}
\alias{ConvertAgeSIA_v2}
\title{Function to convert WHOs SIA age range into something usable}
\usage{
ConvertAgeSIA_v2(sia.age.range, check = TRUE)
}
\arguments{
\item{sia.age.range}{input as AgeGroup from the WHO SIA Data}

\item{check}{logical of whether to print diagnostics of the functions success, default is FALSE}
}
\value{
Returns age range in months
}
\description{
Function to convert WHOs SIA age range into something usable
}
