% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{country_codes.rda}
\format{
A data frame with 208 rows and 5 columns
}
\usage{
data(country_codes)
}
\description{
A data set of country codes
}
\keyword{datasets}
