% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{cov.estimates-class}
\alias{cov.estimates-class}
\title{Generate object of class cov.estimates}
\description{
Generate object of class cov.estimates
}
\section{Slots}{

\describe{
\item{\code{sia.cov}}{numeric.}

\item{\code{sia.min.age}}{numeric.}

\item{\code{sia.max.age}}{numeric.}

\item{\code{MR1.cov}}{numeric.}

\item{\code{MR2.cov}}{numeric.}

\item{\code{inaccessible.prop}}{numeric.}
}}

