% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.country.x.DFE.ID.state.matrix.R
\name{create.country.x.DFE.ID.state.matrix}
\alias{create.country.x.DFE.ID.state.matrix}
\title{Function to create a state matrix at the disease free equilibrium (DFE) from a transition object}
\usage{
create.country.x.DFE.ID.state.matrix(
  uncode,
  tot.pop = NULL,
  tran,
  epi.class.label = c("M", "S", "I", "R", "V"),
  year = 1990,
  use_montagu_demog = FALSE,
  demog_data = NULL
)
}
\arguments{
\item{uncode}{UN country code}

\item{tot.pop}{population by which you can rescale if you want}

\item{tran}{the transition object}

\item{epi.class.label}{xxx ask Amy}

\item{year}{numeric; options are 1950 to 2100}

\item{use_montagu_demog}{logical; if TRUE use montagu demography data}

\item{demog_data}{optional demography data}
}
\value{
a starting DFE vector of S and R individuals
}
\description{
Function to create a state matrix at the disease free equilibrium (DFE) from a transition object
}
