% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.ID.state.matrix.R
\name{create.ID.state.matrix}
\alias{create.ID.state.matrix}
\title{Function creates an ID state matrix with the given number of epidemiologic categories, age classes and labels}
\usage{
create.ID.state.matrix(
  n.age.class,
  n.epi.class = 3,
  epi.class.label = c("S", "I", "R"),
  value = rep(0, n.age.class * n.epi.class)
)
}
\arguments{
\item{n.age.class}{the number of age classes}

\item{n.epi.class}{the number of states in the model}

\item{epi.class.label}{the names of the states}

\item{value}{starting values, defauls to all 0}
}
\value{
an ID.state.matrix object
}
\description{
Function creates an ID state matrix with the given number of epidemiologic categories, age classes and labels
}
