% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.ID.transition.SIR.R
\name{create.ID.transition.SIR}
\alias{create.ID.transition.SIR}
\title{Simple method to create an ID.transition.matrix}
\usage{
create.ID.transition.SIR(
  n.age.class,
  aging.rate,
  survival.rate,
  waifw,
  age.class = numeric(),
  n.epi.class = 3,
  birth.rate = 0,
  introduction.rate = 0,
  exponent = 1,
  frequency.dep = F,
  is.stochastic = F,
  get.births = function(pop, tran) {     return(c(tran@birth.rate, rep(0, length(pop) -
    1))) }
)
}
\arguments{
\item{n.age.class}{the number of age classes}

\item{aging.rate}{the rate which people move to the next age class}

\item{survival.rate}{the percent who survive at each time step}

\item{waifw}{whom acquires infection from whom matrix}

\item{age.class}{the actual age classes defined by upper age in class}

\item{n.epi.class}{the number of states in the model}

\item{birth.rate}{xxx}

\item{introduction.rate}{xxx}

\item{exponent}{xxx}

\item{frequency.dep}{xxx}

\item{is.stochastic}{xxx}

\item{get.births}{xxx}
}
\value{
ID transition matrix
}
\description{
Simple method to create an ID.transition.matrix
}
