% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.ID.transition.SIR.vac.R
\name{create.ID.transition.SIR.vac}
\alias{create.ID.transition.SIR.vac}
\title{Method to create an ID.transtion.SIR.vac object}
\usage{
create.ID.transition.SIR.vac(
  n.age.class,
  aging.rate,
  survival.rate,
  waifw,
  routine.vac = 0,
  routine.vac.age.index = 12,
  time.step = 1/2,
  age.class = numeric(),
  n.epi.class = 3,
  birth.rate = 0,
  introduction.rate = 0,
  exponent = 1,
  frequency.dep = F,
  is.stochastic = F,
  get.births = function(pop, tran) {     return(c(tran@birth.rate, rep(0, length(pop) -
    1))) }
)
}
\arguments{
\item{n.age.class}{xxx}

\item{aging.rate}{xxx}

\item{survival.rate}{xxx}

\item{waifw}{xxx}

\item{routine.vac}{xxx}

\item{routine.vac.age.index}{xxx}

\item{time.step}{xxx}

\item{age.class}{xxx}

\item{n.epi.class}{xxx}

\item{birth.rate}{xxx}

\item{introduction.rate}{xxx}

\item{exponent}{xxx}

\item{frequency.dep}{xxx}

\item{is.stochastic}{xxx}

\item{get.births}{xxx}
}
\value{
transition matrix
}
\description{
Method to create an ID.transtion.SIR.vac object
}
