% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.ID.transition.SIR.vac.SIA.R
\name{create.ID.transition.SIR.vac.SIA}
\alias{create.ID.transition.SIR.vac.SIA}
\title{Method to create an ID.transition.SIR.vac object}
\usage{
create.ID.transition.SIR.vac.SIA(
  n.age.class,
  aging.rate,
  survival.rate,
  waifw,
  routine.vac = 0,
  routine.vac.age.index = 12,
  time.step = 1/2,
  age.class = numeric(),
  n.epi.class = 3,
  birth.rate = 0,
  sia.vac = 0,
  sia.vsucc = new("vsucc.constant", success.rate = 1),
  introduction.rate = 0,
  exponent = 1,
  frequency.dep = F,
  is.stochastic = F,
  get.births = function(pop, tran) {     return(c(tran@birth.rate, rep(0, length(pop) -
    1))) }
)
}
\arguments{
\item{n.age.class}{the number of age classes}

\item{aging.rate}{the rate which people move to the next age class}

\item{survival.rate}{the percent who survive at each time step}

\item{waifw}{whom acquires infection from whom matrix}

\item{routine.vac}{xxx}

\item{routine.vac.age.index}{xxx}

\item{time.step}{xxx}

\item{age.class}{the actual age classes defined by upper age in class}

\item{n.epi.class}{the number of states in the model}

\item{birth.rate}{xxx}

\item{sia.vac}{xxx}

\item{sia.vsucc}{xxx}

\item{introduction.rate}{xxx}

\item{exponent}{xxx}

\item{frequency.dep}{xxx}

\item{is.stochastic}{xxx}

\item{get.births}{xxx}
}
\value{
transition matrix
}
\description{
Method to create an ID.transition.SIR.vac object
}
