% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.surv.prob.over.age.R
\name{create.surv.prob.over.age}
\alias{create.surv.prob.over.age}
\title{Function to pull in Age Specific Death Rates and return survivorship for running out transients}
\usage{
create.surv.prob.over.age(
  age.classes,
  generation.time,
  check = F,
  nMx = NULL,
  year = 1990
)
}
\arguments{
\item{age.classes}{a set of age classes for which tran is being built}

\item{generation.time}{generation time}

\item{check}{xxx}

\item{nMx}{object of age specific death rates over time and mid-age per rate}

\item{year}{year to pull age specific death rates that will be used to run out the transients (should coincide with DFE year)}
}
\value{
returns the age profile of survivorship in units of the generation time
}
\description{
Function to pull in Age Specific Death Rates and return survivorship for running out transients
}
