% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.surv.prob.over.age.time.R
\name{create.surv.prob.over.age.time}
\alias{create.surv.prob.over.age.time}
\title{Function to pull Age Specific Death Rates and return survivorship over time}
\usage{
create.surv.prob.over.age.time(
  age.classes,
  generation.time,
  nMx = NULL,
  nMx.years,
  check = F
)
}
\arguments{
\item{age.classes}{a set of age classes for which tran is being built}

\item{generation.time}{generation time}

\item{nMx}{object of age specific death rates over time and mid-age per rate}

\item{nMx.years}{year to start pulling age specific death rates}

\item{check}{xxx}
}
\value{
the age profile of survivorship in units of the generation time
}
\description{
Function to pull Age Specific Death Rates and return survivorship over time
}
