% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DHS_ageMCV1_Data}
\alias{DHS_ageMCV1_Data}
\title{DHS_ageMCV1_Data.rda}
\format{
A data frame with 268880 rows and 5 columns
}
\usage{
data(DHS_ageMCV1_Data)
}
\description{
A data set
}
\keyword{datasets}
