% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DHS_ageMCV1_Data_AFRO}
\alias{DHS_ageMCV1_Data_AFRO}
\title{DHS_ageMCV1_Data_AFRO.rda}
\format{
A data frame with 25000 rows and 5 columns
}
\usage{
data(DHS_ageMCV1_Data_AFRO)
}
\description{
A data set
}
\keyword{datasets}
