% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtp1_estimates}
\alias{dtp1_estimates}
\title{dtp1_estimates.rda}
\format{
A data frame with 23474 rows and 7 columns
}
\usage{
data(dtp1_estimates)
}
\description{
A data set
}
\keyword{datasets}
