% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EX.Country.part1.setSUS.R
\name{EX.Country.part1.setSUS}
\alias{EX.Country.part1.setSUS}
\title{Function to set up experiment and run the transients out}
\usage{
EX.Country.part1.setSUS(
  uncode,
  generation.time = 0.5,
  age.classes = c(1:240, seq(241, 720, 12)),
  maternal.decay.rt = 0.95,
  exponent = 0.97,
  frequency.dep = TRUE,
  tot.pop = NULL,
  yr.births.per.1000.acrossyears,
  intro.rate = 1e-10,
  targeted.intro = FALSE,
  R0 = 5,
  t.max = 20,
  get.births = NULL,
  seasonal.amp = 0.35,
  flat.WAIFW = FALSE,
  country.specific.WAIFW = FALSE,
  vynnycky.waifw = FALSE,
  vynnycky.waifw.betas = c(2, 1),
  asdr.object,
  year = 1990,
  use_montagu_demog = FALSE,
  routine.vac = 0,
  routine.vac.age.index = 12,
  demog_data = NULL,
  starting.prop.immune = c(0.75, 0.75, 0.75, 0.75),
  starting.prop.immune.ages.in.months = c(9, 12, 24, 240)
)
}
\arguments{
\item{uncode}{UN country code}

\item{generation.time}{the desired generation time in months}

\item{age.classes}{vector; the upper limit of the age classes we are interested in months}

\item{maternal.decay.rt}{rate of maternal decay of immunity to rubella}

\item{exponent}{numeric; exponent for the infected}

\item{frequency.dep}{xxx}

\item{tot.pop}{numeric; population you want to scale you whole experiment by at the beginning, NULL or numeric}

\item{yr.births.per.1000.acrossyears}{vector of length t.max; crude birth rate per 1000 per year}

\item{intro.rate}{numeric; rate at which new infections are introduced}

\item{targeted.intro}{logical; FALSE means space introduction out over all age classes, TRUE means to concentrate introductions}

\item{R0}{numeric; basic reproductive number of rubella assumed, used to rescale the waifw}

\item{t.max}{numeric; time in year that plan to run out transients}

\item{get.births}{vector of births with length of 5*n.age.classes}

\item{seasonal.amp}{numeric; amplitude of the seasonal cosine function}

\item{flat.WAIFW}{logical; if FALSE then default waifw is polymod GB}

\item{country.specific.WAIFW}{logical; if true then overrides polymod or flat waifw}

\item{vynnycky.waifw}{logical; true if want emilia's waifw <13 and 15-50 (beta_young, beta_old, and 0.7*beta_old)}

\item{vynnycky.waifw.betas}{vector of two betas (beta_young and beta_old)}

\item{asdr.object}{nMx object with rate and mid-age (age specific death rates)}

\item{year}{numeric; year to pull DFE demography (population and age structure)}

\item{use_montagu_demog}{logical; T = use montagu demography or F = use UNPD demography}

\item{routine.vac}{xxx}

\item{routine.vac.age.index}{xxx}

\item{demog_data}{optional demography data for use if use_montagu_demog==T}

\item{starting.prop.immune}{vector - starting proportion immune associated with starting.prop.immune.ages.in.months}

\item{starting.prop.immune.ages.in.months}{vector - the ages (in months) associated with starting.prop.immune}
}
\value{
experiment result
}
\description{
Function to set up experiment and run the transients out
}
