% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EX.Country.part2.R
\name{EX.Country.part2}
\alias{EX.Country.part2}
\title{Function to Run the Simulations/Experiments}
\usage{
EX.Country.part2(
  uncode,
  generation.time = 0.5,
  pop.rescale = NULL,
  pop.time = 4,
  is.stochastic = FALSE,
  get.births = NULL,
  t.max = 40,
  rescale.WAIFW = F,
  yr.births.per.1000.acrossyears,
  asdr.object,
  year = 1990,
  EXt0 = EXt0,
  time.specific.MR1cov,
  time.specific.MR2cov,
  time.specific.SIAcov,
  time.specific.min.age.MR1,
  time.specific.max.age.MR1,
  time.specific.min.age.MR2,
  time.specific.max.age.MR2,
  time.specific.min.age.SIA,
  time.specific.max.age.SIA,
  obj.vcdf.MR1 = get.vcdf.normal(6, 12),
  obj.vcdf.MR2 = get.vcdf.normal(15, 21),
  obj.prob.vsucc = pvacsuccess(1:(20 * 12), get.boulianne.vsucc()),
  sia.timing.in.year = (3/12),
  MR1MR2correlation = FALSE,
  MR1SIAcorrelation = 0,
  MR2SIAcorrelation = 0,
  SIAinacc = FALSE,
  prop.inacc = NULL,
  SIAinefficient = FALSE,
  intro.rate
)
}
\arguments{
\item{uncode}{UN country code}

\item{generation.time}{generation time in months}

\item{pop.rescale}{numeric; population by which you want to rescale at pop.time}

\item{pop.time}{numeric; time in YEARS that you want to rescale the population by}

\item{is.stochastic}{xxx}

\item{get.births}{vector of births with length of 5*n.age.classes}

\item{t.max}{numeric; time in year that plan to run the experiment}

\item{rescale.WAIFW}{xxx}

\item{yr.births.per.1000.acrossyears}{vector of length t.max; crude birth rate per 1000 per year}

\item{asdr.object}{nMx object; with rate and mid-age (age specific death rates)}

\item{year}{numeric; year that pulled DFE demography (population and age structure) from part1}

\item{EXt0}{object from EX.Country.part1()}

\item{time.specific.MR1cov}{numeric; coverage in proportion by year}

\item{time.specific.MR2cov}{numeric; coverage in proportion by year}

\item{time.specific.SIAcov}{xxx}

\item{time.specific.min.age.MR1}{xxx}

\item{time.specific.max.age.MR1}{xxx}

\item{time.specific.min.age.MR2}{xxx}

\item{time.specific.max.age.MR2}{xxx}

\item{time.specific.min.age.SIA}{xxx}

\item{time.specific.max.age.SIA}{xxx}

\item{obj.vcdf.MR1}{xxx}

\item{obj.vcdf.MR2}{xxx}

\item{obj.prob.vsucc}{xxx}

\item{sia.timing.in.year}{xxx}

\item{MR1MR2correlation}{xxx}

\item{MR1SIAcorrelation}{xxx}

\item{MR2SIAcorrelation}{xxx}

\item{SIAinacc}{xxx}

\item{prop.inacc}{xxx}

\item{SIAinefficient}{xxx}

\item{intro.rate}{- vector for each time point (this is the rate per group)}
}
\value{
experiment results
}
\description{
Function to Run the Simulations/Experiments
}
