% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{experiment-class}
\alias{experiment-class}
\title{Class to define an experiment.}
\description{
Class to define an experiment.
}
\section{Slots}{

\describe{
\item{\code{name}}{character.}

\item{\code{R0}}{numeric.}

\item{\code{state.t0}}{ID.state.matrix.}

\item{\code{t.min}}{numeric.}

\item{\code{t.max}}{numeric.}

\item{\code{t0.doy}}{numeric.}

\item{\code{step.size}}{numeric.}

\item{\code{trans}}{ID.transition.SIR.}

\item{\code{season.obj}}{seasonal.force.}

\item{\code{births.per.1000.each.timestep}}{numeric.}

\item{\code{description}}{character.}
}}

