% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{experiment.result-class}
\alias{experiment.result-class}
\title{Holds the results of an experiment along with the experiment definition}
\description{
Holds the results of an experiment along with the experiment definition
}
\section{Slots}{

\describe{
\item{\code{experiment.def}}{experiment.}

\item{\code{result}}{sim.results.SIR.}
}}

