% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{experiment.SIAtrigger-class}
\alias{experiment.SIAtrigger-class}
\title{Class defines an experiment with SIAs or opportunity for triggered SIAs}
\description{
Class defines an experiment with SIAs or opportunity for triggered SIAs
}
\section{Slots}{

\describe{
\item{\code{trans}}{ID.transition.SIR.vac.}

\item{\code{sia.obj}}{matrix.}

\item{\code{trigger.list}}{list.}
}}

