% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{experiment.updatedemog-class}
\alias{experiment.updatedemog-class}
\title{Class for an experiment with birth and death rates updated each time step and population rescaling}
\description{
Class for an experiment with birth and death rates updated each time step and population rescaling
}
\section{Slots}{

\describe{
\item{\code{trans}}{ID.transition.SIR.vac.}

\item{\code{surv.each.timestep}}{matrix.}

\item{\code{pop.rescale.each.timestep}}{numeric.}

\item{\code{maternal.obj}}{maternal.exp.decay.}
}}

