% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{experiment.updatedemog.vaccinationchange-class}
\alias{experiment.updatedemog.vaccinationchange-class}
\title{Class for an experiment with changing vaccination coverage over time}
\description{
Class for an experiment with changing vaccination coverage over time
}
\section{Slots}{

\describe{
\item{\code{time.specific.MR1cov}}{vector.}

\item{\code{time.specific.MR2cov}}{vector.}

\item{\code{time.specific.SIAcov}}{vector.}

\item{\code{time.specific.min.age.MR1}}{vector.}

\item{\code{time.specific.max.age.MR1}}{vector.}

\item{\code{time.specific.min.age.MR2}}{vector.}

\item{\code{time.specific.max.age.MR2}}{vector.}

\item{\code{time.specific.min.age.SIA}}{vector.}

\item{\code{time.specific.max.age.SIA}}{vector.}

\item{\code{obj.vcdf.MR1}}{vaccine.cdf.byage.}

\item{\code{obj.vcdf.MR2}}{vaccine.cdf.byage.}

\item{\code{obj.prob.vsucc}}{prob.vsucc.byage.}

\item{\code{sia.timing.in.year}}{vector.}
}}

