% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{experiment.updatedemog.vaccinationchange.vaccinationcorrelation-class}
\alias{experiment.updatedemog.vaccinationchange.vaccinationcorrelation-class}
\title{Class for an experiment with RI-SIA dose correlation}
\description{
Class for an experiment with RI-SIA dose correlation
}
\section{Slots}{

\describe{
\item{\code{MR1MR2correlation}}{logical.}

\item{\code{MR1SIAcorrelation}}{numeric.}

\item{\code{MR2SIAcorrelation}}{numeric.}
}}

