% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{experiment.updatedemog.vaccinationchange.vaccinationcorrelation.outbreakresponse-class}
\alias{experiment.updatedemog.vaccinationchange.vaccinationcorrelation.outbreakresponse-class}
\title{Class for an outbreak response experiment}
\description{
Class for an outbreak response experiment
}
\section{Slots}{

\describe{
\item{\code{or.total.delay}}{numeric. Total time steps from end of surveillance window to response.}

\item{\code{or.trigger.window}}{numeric. Number of time steps to sum I over for trigger metric.}

\item{\code{or.threshold.value}}{numeric. Threshold for trigger metric.}

\item{\code{or.threshold.type}}{character. "count" or "incidence".}

\item{\code{or.trigger.age.lower}}{numeric. Lower age bound for surveillance trigger (months); NA = all ages.}

\item{\code{or.trigger.age.upper}}{numeric. Upper age bound for surveillance trigger (months); NA = all ages.}

\item{\code{or.vacc.age.lower}}{numeric. Lower age bound for OBR vaccination campaign (months).}

\item{\code{or.vacc.age.upper}}{numeric. Upper age bound for OBR vaccination campaign (months).}

\item{\code{or.vacc.coverage}}{numeric. OBR campaign coverage in [0,1].}

\item{\code{or.min.interval}}{numeric. Minimum time steps between successive OBR triggers.}
}}

