% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{experiment.updatedemog.vaccinationchange.vaccinationlimitations-class}
\alias{experiment.updatedemog.vaccinationchange.vaccinationlimitations-class}
\title{Class for an experiment}
\description{
Class for an experiment
}
\section{Slots}{

\describe{
\item{\code{MR1MR2correlation}}{logical.}

\item{\code{MR1SIAcorrelation}}{logical.}

\item{\code{MR2SIAcorrelation}}{logical.}

\item{\code{SIAinefficient}}{logical.}

\item{\code{SIAinacc}}{logical.}

\item{\code{prop.inacc}}{numeric.}
}}

