% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{experiment.updatedemog.vaccinationchange.vaccinationlimitations.spatial-class}
\alias{experiment.updatedemog.vaccinationchange.vaccinationlimitations.spatial-class}
\title{Class for a spatial vaccination experiment updating demography and with changing vaccination coverage over time, with vaccination limitations}
\description{
Class for a spatial vaccination experiment updating demography and with changing vaccination coverage over time, with vaccination limitations
}
\section{Slots}{

\describe{
\item{\code{trans}}{ID.transition.MSIRV.space.}
}}

