% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractAgeSpecificSurvivalMatrix.R
\name{ExtractAgeSpecificSurvivalMatrix}
\alias{ExtractAgeSpecificSurvivalMatrix}
\title{Function to produce age survival matrix}
\usage{
ExtractAgeSpecificSurvivalMatrix(tmp.trans, maternal.obj, surv.at.timestep.t)
}
\arguments{
\item{tmp.trans}{transitions object of class ID.transition.MSIRV}

\item{maternal.obj}{maternal object of class maternal.exp.decay}

\item{surv.at.timestep.t}{the estimated age-specific survival rates at time t}
}
\value{
the age survival matrix per time t
}
\description{
Function to produce age survival matrix
}
