% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.vsucc.Logistic.R
\name{fit.vsucc.Logistic}
\alias{fit.vsucc.Logistic}
\title{Fits a vsucc.Logistic object to some data using logistic regression}
\usage{
fit.vsucc.Logistic(data, full.efficacy)
}
\arguments{
\item{data}{a data frame of age (in months), number of success and total tests}

\item{full.efficacy}{the maximum efficacy we wish to allow}
}
\value{
a vsucc.Logistic object
}
\description{
Fits a vsucc.Logistic object to some data using logistic regression
}
