% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.boulianne.vsucc.R
\name{get.boulianne.vsucc}
\alias{get.boulianne.vsucc}
\title{Get the Boulianne vaccine success object (rubella)}
\usage{
get.boulianne.vsucc(saturation = 0.97)
}
\value{
xxx
}
\description{
Get the Boulianne vaccine success object (rubella)
}
