% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.CountryX.Starting.Pop.MSIRV.R
\name{Get.CountryX.Starting.Pop.MSIRV}
\alias{Get.CountryX.Starting.Pop.MSIRV}
\title{Function to get starting state and transition object}
\usage{
Get.CountryX.Starting.Pop.MSIRV(
  uncode,
  generation.time = 0.5,
  age.classes = c(1:240, seq(241, 720, 12)),
  maternal.decay.rt = 0.95,
  exponent = 0.97,
  frequency.dep = TRUE,
  is.stochastic = FALSE,
  tot.pop = NULL,
  yr.births.per.1000,
  intro.rate = intro.rate,
  flat.WAIFW = FALSE,
  asdr.object = NULL,
  targeted.intro = FALSE,
  year = 1990,
  get.births,
  use_montagu_demog = TRUE,
  routine.vac = 0,
  routine.vac.age.index = 12,
  demog_data = NULL
)
}
\arguments{
\item{uncode}{UN country code}

\item{generation.time}{the desired generation time in months}

\item{age.classes}{vector; the upper limit of the age classes we are interested in months}

\item{maternal.decay.rt}{rate of maternal decay of immunity to rubella}

\item{exponent}{numeric; exponent for the infected}

\item{frequency.dep}{logical; TRUE}

\item{is.stochastic}{logical; FALSE}

\item{tot.pop}{numeric; population you want to scale you whole experiment by at the beginning, NULL or numeric}

\item{yr.births.per.1000}{numeric; crude birth rate per year per 1000}

\item{intro.rate}{xxx}

\item{flat.WAIFW}{logical; if F then default waifw is polymod GB}

\item{asdr.object}{nMx object with rate and mid-age (age specific death rates)}

\item{targeted.intro}{logical; FALSE means space introduction out over all age classes, TRUE means to concentrate introductions}

\item{year}{numeric; year to pull country DFE demography (age structure and population size)}

\item{get.births}{xxx}

\item{use_montagu_demog}{logical; if TRUE use montagu demography data}

\item{routine.vac}{xxx}

\item{routine.vac.age.index}{xxx}

\item{demog_data}{optional demography data for use if use_montagu_demog = TRUE}
}
\value{
starting state and transition object
}
\description{
Function to get starting state and transition object
}
