% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.diagonal.WAIFW.R
\name{get.diagonal.WAIFW}
\alias{get.diagonal.WAIFW}
\title{Make a diagonal WAIFW matrix}
\usage{
get.diagonal.WAIFW(age.class.boundries = (1:120/12))
}
\arguments{
\item{age.class.boundries}{the upper age limit for each age class in years}
}
\value{
a diagonal WAIFW matrix
}
\description{
Make a diagonal WAIFW matrix
}
