% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.flat.WAIFW.R
\name{get.flat.WAIFW}
\alias{get.flat.WAIFW}
\title{Make a flat WAIFW matrix}
\usage{
get.flat.WAIFW(age.class.boundries = (1:120/12))
}
\arguments{
\item{age.class.boundries}{the upper age limit for each age class in years}
}
\value{
a flat WAIFW matrix
}
\description{
Make a flat WAIFW matrix
}
