% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.MR1cdf.survival.R
\name{get.MR1cdf.survival}
\alias{get.MR1cdf.survival}
\title{Function to find the non-scaled MR1 CDF (uses most recent DHS data as of 2019)}
\usage{
get.MR1cdf.survival(uncode, min.age = 1, max.age = 24)
}
\arguments{
\item{uncode}{numeric; country UN code}

\item{min.age}{numeric; age in months}

\item{max.age}{numeric; age in months}
}
\value{
CDF vaccinated across age in months
}
\description{
Function to find the non-scaled MR1 CDF (uses most recent DHS data as of 2019)
}
