% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.polymod.WAIFW.R
\name{get.polymod.WAIFW}
\alias{get.polymod.WAIFW}
\title{Make a WAIFW matrix based on Polymod (ignoring rescaling by population size that might be sensible - just use raw contacts)}
\usage{
get.polymod.WAIFW(
  age.class.boundries = (1:90),
  country = "GB",
  bandwidth = c(3, 3),
  do.touch = FALSE
)
}
\arguments{
\item{age.class.boundries}{the upper age limit for each age class in YEARS}

\item{country}{levels possible "all", "IT" "DE" "LU" "NL" "PL" "GB" "FI" "BE", default is "GB"}

\item{bandwidth}{desired smooth bandwidth - default=c(3,3)}

\item{do.touch}{do you want to only include contacts involving touching? default is FALSE}
}
\value{
a WAIFW matrix based on the Polymod results from chosen location with row and colnames indicating age classes
}
\description{
Make a WAIFW matrix based on Polymod (ignoring rescaling by population size that might be sensible - just use raw contacts)
}
