% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.seasonal.mult.R
\docType{methods}
\name{get.seasonal.mult}
\alias{get.seasonal.mult}
\alias{get.seasonal.mult,numeric,seasonal.cosine-method}
\alias{get.seasonal.mult,numeric,seasonal.periodic.bspline-method}
\alias{get.seasonal.mult,numeric,seasonal.piecewise.constant-method}
\alias{get.seasonal.mult,numeric,seasonal.age.specific-method}
\alias{get.seasonal.mult,numeric,seasonal.age.specific.piecewise.constant-method}
\title{Get seasonal adjustment for times based on a seasonality object}
\usage{
get.seasonal.mult(t.in.yr, season.obj)

\S4method{get.seasonal.mult}{numeric,seasonal.cosine}(t.in.yr, season.obj)

\S4method{get.seasonal.mult}{numeric,seasonal.periodic.bspline}(t.in.yr, season.obj)

\S4method{get.seasonal.mult}{numeric,seasonal.piecewise.constant}(t.in.yr, season.obj)

\S4method{get.seasonal.mult}{numeric,seasonal.age.specific}(t.in.yr, season.obj)

\S4method{get.seasonal.mult}{numeric,seasonal.age.specific.piecewise.constant}(t.in.yr, season.obj)
}
\arguments{
\item{t.in.yr}{time of the year in fraction of 1 year}

\item{season.obj}{an object encoding the seasonality}
}
\value{
the seasonal multiplier, or series thereof
}
\description{
Get seasonal adjustment for times based on a seasonality object
}
