% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.sia.time.age.specific.R
\name{get.sia.time.age.specific}
\alias{get.sia.time.age.specific}
\title{Get SIA time age-specific}
\usage{
get.sia.time.age.specific(
  age.classes = c(1:240, seq(252, 1212, 12)),
  time.specific.SIAcov = c(rep(0, 29), 0.5, rep(0, 19), 0.7),
  age.min.sia = rep(12, 50),
  age.max.sia = rep(60, 50),
  obj.prob.vsucc = pvacsuccess(1:240, new("vsucc.constant", success.rate = 0.85))
)
}
\arguments{
\item{age.classes}{xxx}

\item{time.specific.SIAcov}{xxx}

\item{age.min.sia}{xxx}

\item{age.max.sia}{xxx}

\item{obj.prob.vsucc}{xxx}
}
\value{
xxx
}
\description{
Get SIA time age-specific
}
