% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.smooth.WAIFW.R
\name{get.smooth.WAIFW}
\alias{get.smooth.WAIFW}
\title{Make a parametric smooth WAIFW from Farrington, J. Amer. Stat. Assoc.; 2005, 100 p370}
\usage{
get.smooth.WAIFW(
  age.class.boundries = (1:120/12),
  mu = 12.71,
  sig = 0.69,
  gam = 0.17,
  delta = 0
)
}
\arguments{
\item{age.class.boundries}{the upper age limit for each age class in years}

\item{mu}{age of highest contact increases with mu}

\item{sig}{width around equal age diagonal increases with gam}

\item{gam}{decreases strength in other diagonal (shrinks high trans int)}

\item{delta}{background homogeneous contact rate}
}
\value{
a smooth WAIFW matrix
}
\description{
Make a parametric smooth WAIFW from Farrington, J. Amer. Stat. Assoc.; 2005, 100 p370
}
