% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.vcdf.normal.R
\name{get.vcdf.normal}
\alias{get.vcdf.normal}
\title{Function to find the non-scaled vaccination CDF (assuming normal pdf)}
\usage{
get.vcdf.normal(min.age = 25, max.age = 36)
}
\arguments{
\item{min.age}{numeric; age in months}

\item{max.age}{numeric; age in months}
}
\value{
CDF vaccinated across age in months
}
\description{
Function to find the non-scaled vaccination CDF (assuming normal pdf)
}
