% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.vynnycky.WAIFW.R
\name{get.vynnycky.WAIFW}
\alias{get.vynnycky.WAIFW}
\title{Make a WAIFW matrix to mimic Emilia's - currently data-free}
\usage{
get.vynnycky.WAIFW(age.class.boundries = (1:120/12), beta_young, beta_old)
}
\arguments{
\item{age.class.boundries}{the upper age limit for each age class in years}

\item{beta_young}{beta on young individuls (<=13 years old)}

\item{beta_old}{beta on old individuals (>13 years old)}
}
\value{
a WAIFW matrix that mimics Emilia's
}
\description{
Make a WAIFW matrix to mimic Emilia's - currently data-free
}
