% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompartmentTotals.R
\docType{methods}
\name{getCompartmentTotals}
\alias{getCompartmentTotals}
\alias{getCompartmentTotals,sim.results.SIR-method}
\alias{getCompartmentTotals,sim.results.MSIRV-method}
\title{Get the total number of susceptible, infected, and recovered from sim.results.SIR}
\usage{
getCompartmentTotals(sim.res, ...)

\S4method{getCompartmentTotals}{sim.results.SIR}(sim.res)

\S4method{getCompartmentTotals}{sim.results.MSIRV}(sim.res)
}
\arguments{
\item{sim.res}{sim result object}

\item{...}{ignored}
}
\value{
a sim result object with no age classes
}
\description{
Get the total number of susceptible, infected, and recovered from sim.results.SIR
}
