% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCRScases.byage.per.stochastic.rate.R
\name{getCRScases.byage.per.stochastic.rate}
\alias{getCRScases.byage.per.stochastic.rate}
\title{Function to get the total number of CRS pregnancies by age}
\usage{
getCRScases.byage.per.stochastic.rate(
  sim.res,
  crs_rate_gestational_age_nbiweeks = c(6, 8),
  crs_rate = c(0.4, 0.15),
  trans,
  no.gens.in.year,
  births,
  asfr.over.time,
  repro.age.sex.dist.over.time
)
}
\arguments{
\item{sim.res}{a sim.result.MSIRV object}

\item{crs_rate_gestational_age_nbiweeks}{vector; gestational ages (in biweeks) that coincides with crs_rate}

\item{crs_rate}{vector; probability crs case given rubella infection in pregnancy and live birth that coincides with crs_rate_gestational_age_nbiweeks}

\item{trans}{xxx}

\item{no.gens.in.year}{xxx}

\item{births}{vector; number of births per time period taken from the simulation}

\item{asfr.over.time}{matrix; age-specific fertility rate in every age class and year (nrow=7, ncol=length of sim in years)}

\item{repro.age.sex.dist.over.time}{xxx}
}
\value{
an object with age classes set by fertility age classes
}
\description{
Function to get the total number of CRS pregnancies by age
}
