% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCumRiskTrimester.R
\docType{methods}
\name{getCumRiskTrimester}
\alias{getCumRiskTrimester}
\alias{getCumRiskTrimester,sim.results.MSIRV,numeric-method}
\title{Get cumulative probability of infection during pregnancy}
\usage{
getCumRiskTrimester(sim.res, nbiweeks, ...)

\S4method{getCumRiskTrimester}{sim.results.MSIRV,numeric}(sim.res, nbiweeks, ...)
}
\arguments{
\item{sim.res}{a sim.result.MSIRV object}

\item{nbiweeks}{chosen number of biweeks of risk}

\item{...}{ignored}
}
\value{
an object with age classes set by fertility age classes
}
\description{
Calculate the cumulative probability of getting infected
over a period of n weeks from a sim.results.MSIRV object
to account for the fact that baby can get CRS if
mother gets sick during first trimester (1st three months)
}
