% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDemography.R
\name{getDemography}
\alias{getDemography}
\title{Function to get population size, age structure, fertility and death rates from 1950 to 2105}
\usage{
getDemography(uncode = NA)
}
\arguments{
\item{uncode}{UN country code}
}
\value{
demography for each country
}
\description{
Function to get population size, age structure, fertility and death rates from 1950 to 2105
}
