% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDemography.wpp2024.R
\name{getDemography.wpp2024}
\alias{getDemography.wpp2024}
\title{Function to get population size, age structure, and fertility from 1950 to 2105
No differences from wpp2024 except updated package}
\usage{
getDemography.wpp2024(
  uncode = NA,
  age.classes = c(1:101),
  if.missing.use.region = F
)
}
\arguments{
\item{uncode}{UN country code}

\item{age.classes}{age classes in years}

\item{if.missing.use.region}{logical; c(MHL, XK, TUV)}
}
\value{
demography for each country
}
\description{
Function to get population size, age structure, and fertility from 1950 to 2105
No differences from wpp2024 except updated package
}
