% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMontaguCoverage_202110gavi_v3.R
\name{getMontaguCoverage_202110gavi_v3}
\alias{getMontaguCoverage_202110gavi_v3}
\title{Function to pull in vaccination data from Montagu 202110sgavi_v3 touchstone vaccine scenarios}
\usage{
getMontaguCoverage_202110gavi_v3(iso3code)
}
\arguments{
\item{iso3code}{iso3code}
}
\value{
age, year, and coverage for SIA, RCV1, and RCV2 for each vaccination scenario
}
\description{
Function to pull in vaccination data from Montagu 202110sgavi_v3 touchstone vaccine scenarios
}
