% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNumber.per.AgeGroup.R
\name{GetNumber.per.AgeGroup}
\alias{GetNumber.per.AgeGroup}
\title{Function to get number of individuals per age group}
\usage{
GetNumber.per.AgeGroup(trans, state, tval = 1)
}
\arguments{
\item{trans}{transition object}

\item{state}{vector of state of the population at one time point}

\item{tval}{xxx}
}
\value{
vector size n.age.classes with the number of individuals in each age
}
\description{
Function to get number of individuals per age group
}
