% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNumberRubellaSusceptible.per.Year.R
\name{GetNumberRubellaSusceptible.per.Year}
\alias{GetNumberRubellaSusceptible.per.Year}
\title{Function to get age-specific number rubella susceptible per year}
\usage{
GetNumberRubellaSusceptible.per.Year(res, trans, epi.state, no.gens.in.year)
}
\arguments{
\item{res}{result data matrix}

\item{trans}{transition object}

\item{epi.state}{epidemiological state indexes}

\item{no.gens.in.year}{xxx}
}
\value{
matrix of seroprevalence by age (in years) per year
}
\description{
Function to get age-specific number rubella susceptible per year
}
