% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutput_201910gavi_v4.R
\name{getOutput_201910gavi_v4}
\alias{getOutput_201910gavi_v4}
\title{Function to get the VIMC output needed from each result object}
\usage{
getOutput_201910gavi_v4(
  sim.res,
  t.max,
  setup,
  year,
  crs_rate_gestational_age_nbiweeks,
  crs_rate_agespecific,
  crs_rate_overall,
  fetal_death_rate,
  infant_death_rate,
  scenario,
  run.number,
  R0
)
}
\arguments{
\item{sim.res}{a sim.result.MSIRV object}

\item{t.max}{numeric; the t.max used to create sim.res}

\item{setup}{country specific object used to create sim.res}

\item{year}{numeric; used to create sim.res}

\item{crs_rate_gestational_age_nbiweeks}{vector; gestational ages (in biweeks) that coincides with crs_rate_agespecific}

\item{crs_rate_agespecific}{vector; probability crs case given rubella infection in pregnancy and live birth that coincides with crs_rate_gestational_age_nbiweeks}

\item{crs_rate_overall}{numeric; average CRS rate over age groups per crs_rate_gestational_age_nbiweeks}

\item{fetal_death_rate}{numeric; prob fetal death given rubella infection in pregnancy}

\item{infant_death_rate}{numeric; prob infant death given rubella infection in pregnancy}

\item{scenario}{xxx}

\item{run.number}{xxx}

\item{R0}{xxx}
}
\value{
a large list of matrices by age and time in biweeks (population, rubella cases, rubella seroprevalence, CRS cases, CRS deaths, CRS dalys, CRS rate)
}
\description{
Function to get the VIMC output needed from each result object
}
