% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetProp.per.Year.R
\name{GetPop.per.Year}
\alias{GetPop.per.Year}
\title{Function to get age-specific population size at mid-year}
\usage{
GetPop.per.Year(res, trans, no.gens.in.year)
}
\arguments{
\item{res}{result data matrix}

\item{trans}{transition object}

\item{no.gens.in.year}{xxx}
}
\value{
matrix of pop size by age (in years) per year
}
\description{
Function to get age-specific population size at mid-year
}
