% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRubellaCases.per.Year.R
\name{GetRubellaCases.per.Year}
\alias{GetRubellaCases.per.Year}
\title{Function to get age-specific rubella cases size per year}
\usage{
GetRubellaCases.per.Year(res, trans, epi.state, no.gens.in.year)
}
\arguments{
\item{res}{result data matrix}

\item{trans}{transition object}

\item{epi.state}{epidemiological state indexes}

\item{no.gens.in.year}{xxx}
}
\value{
matrix of pop size by age (in years) per year
}
\description{
Function to get age-specific rubella cases size per year
}
