% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRubellaSeroprevalence.per.Year.R
\name{GetRubellaSeroprevalence.per.Year}
\alias{GetRubellaSeroprevalence.per.Year}
\title{Function to get age-specific rubella seroprevalence per year}
\usage{
GetRubellaSeroprevalence.per.Year(
  res,
  trans,
  epi.state,
  no.gens.in.year,
  age0is9to12monly = FALSE
)
}
\arguments{
\item{res}{result data matrix}

\item{trans}{transition object}

\item{epi.state}{epidemiological state indexes}

\item{no.gens.in.year}{xxx}
}
\value{
matrix of seroprevalence by age (in years) per year
}
\description{
Function to get age-specific rubella seroprevalence per year
}
