% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSumInYear.R
\name{GetSumInYear}
\alias{GetSumInYear}
\title{Sums over no.gens.in.year to get number per year}
\usage{
GetSumInYear(mat, no.gens.in.year)
}
\arguments{
\item{mat}{matrix with ncol equal to no.gens.year*year}

\item{no.gens.in.year}{xxx}
}
\value{
xxx
}
\description{
Sums over no.gens.in.year to get number per year
}
