% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWHO.measles.SIAs.R
\name{getWHO.measles.SIAs}
\alias{getWHO.measles.SIAs}
\title{Function to convert measles SIAs into something usable}
\usage{
getWHO.measles.SIAs(iso3code, uncode, upper.limit = 0.97)
}
\arguments{
\item{iso3code}{iso3code}

\item{uncode}{UN country code}

\item{upper.limit}{upper coverage limit as proportion}
}
\value{
age range, year, and coverage of each SIA
}
\description{
Function to convert measles SIAs into something usable
}
