% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWHO.rubella.SIAs.Nov2023.R
\name{getWHO.rubella.SIAs.Nov2023}
\alias{getWHO.rubella.SIAs.Nov2023}
\title{Function to convert rubella SIAs into something usable}
\usage{
getWHO.rubella.SIAs.Nov2023(iso3code, uncode, upper.limit = 0.97, demog)
}
\arguments{
\item{iso3code}{ido3code}

\item{uncode}{UN country code}

\item{upper.limit}{upper coverage limit as proportion}

\item{demog}{demography from setupCountry.Nov.2023()}
}
\value{
age range, year, and coverage of each SIA
}
\description{
Function to convert rubella SIAs into something usable
}
