% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWHO.SIAs.Sept2026.R
\name{getWHO.SIAs.Sep2026}
\alias{getWHO.SIAs.Sep2026}
\title{Function to convert measles SIAs into something usable}
\usage{
getWHO.SIAs.Sep2026(
  iso3code,
  uncode,
  upper.limit = 0.97,
  demog,
  rubella.SIAs = FALSE
)
}
\arguments{
\item{iso3code}{iso3code}

\item{uncode}{UN country code}

\item{upper.limit}{upper coverage limit as proportion}

\item{demog}{demography from setupCountry.Jun2026()}

\item{rubella.SIAs}{boolean - default FALSE means pulls measles SIAs, if TRUE then pulls rubella SIAs}
}
\value{
age range, year, and coverage of each SIA
}
\description{
Function to convert measles SIAs into something usable
}
