% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{ID.state.matrix-class}
\alias{ID.state.matrix-class}
\title{Generate object of class ID.state.matrix}
\description{
Generate object of class ID.state.matrix
}
\section{Slots}{

\describe{
\item{\code{n.epi.class}}{numeric.}

\item{\code{epi.class}}{numeric.}

\item{\code{epi.class.label}}{character.}

\item{\code{n.age.class}}{numeric.}
}}

