% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{ID.transition.MSIRV.SIA-class}
\alias{ID.transition.MSIRV.SIA-class}
\title{Generates an object of class ID.transitions.MSIRV.SIA}
\description{
Class to represent transitions in the context of vaccination
and maternal antibodies.
}
\section{Slots}{

\describe{
\item{\code{m.inds}}{numeric}

\item{\code{v.inds}}{numeric}
}}

