% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{ID.transition.MSIRV-class}
\alias{ID.transition.MSIRV-class}
\title{Generate object of class ID.transition.MSIRV}
\description{
Class to represent transitions in the context of vaccination
and maternal antibodies.
}
\section{Slots}{

\describe{
\item{\code{m.inds}}{numeric.}

\item{\code{v.inds}}{numeric.}
}}

