% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{ID.transition.MSIRV.space-class}
\alias{ID.transition.MSIRV.space-class}
\title{Class to represent transitions in the context of vaccination and maternal antibodies with multiple locations}
\description{
Class to represent transitions in the context of vaccination and maternal antibodies with multiple locations
}
\section{Slots}{

\describe{
\item{\code{n.subpops}}{numeric}

\item{\code{subpop.class.label}}{numeric}

\item{\code{coupling}}{matrix}
}}

