% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{ID.transition.SIR-class}
\alias{ID.transition.SIR-class}
\title{Generates an object of class ID.transition.SIR}
\description{
Class definition for transition matrix. This matrix knows how to shift the
ID.state.matrix from one time step to another, and contains all the fun stuff
we need to do the transition (i.e., the matrix)
}
\section{Slots}{

\describe{
\item{\code{n.epi.class}}{numeric.}

\item{\code{epi.class}}{numeric.}

\item{\code{epi.class.label}}{character.}

\item{\code{s.inds}}{numeric.}

\item{\code{i.inds}}{numeric.}

\item{\code{r.inds}}{numeric.}

\item{\code{n.age.class}}{numeric.}

\item{\code{age.class}}{numeric.}

\item{\code{aging.rate}}{numeric.}

\item{\code{survival.rate}}{numeric.}

\item{\code{birth.rate}}{numeric.}

\item{\code{waifw}}{matrix.}

\item{\code{age.surv.matrix}}{matrix.}

\item{\code{introduction.rate}}{numeric.}

\item{\code{exponent}}{numeric.}

\item{\code{frequency.dep}}{logical.}

\item{\code{is.stochastic}}{logical.}

\item{\code{get.births}}{function.}
}}

