% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{ID.transition.SIR.vac-class}
\alias{ID.transition.SIR.vac-class}
\title{Generates object of class ID.transition.SIR.vac}
\description{
Class to represent transitions in the context of vaccination
}
\section{Slots}{

\describe{
\item{\code{vac.per}}{vacc.per.time.step.by.age.}
}}

