% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{ID.transition.SIR.vac.SIA-class}
\alias{ID.transition.SIR.vac.SIA-class}
\title{Generates object of class ID.transition.SIR.vac.SIA}
\description{
Generates object of class ID.transition.SIR.vac.SIA
}
\section{Slots}{

\describe{
\item{\code{sia.vac}}{numeric.}

\item{\code{sia.vsucc}}{vsucc.}
}}

