% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{maternal.exp.decay-class}
\alias{maternal.exp.decay-class}
\title{Class for maternal immunity}
\description{
Assumes that protection is the same as exponential decay
}
\section{Slots}{

\describe{
\item{\code{decay.rt}}{numeric.}
}}

