% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{maternal.thresh-class}
\alias{maternal.thresh-class}
\title{Class for maternal immunity threshold}
\description{
Assume that there is some threshold age at which
you are 100% protected if this age or younger
}
\section{Slots}{

\describe{
\item{\code{thresh}}{numeric.}
}}

