% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next.ID.state.R
\docType{methods}
\name{next.ID.state}
\alias{next.ID.state}
\alias{next.ID.state,ID.state.matrix,ID.transition.SIR-method}
\alias{next.ID.state,ID.state.matrix,ID.transition.SIR.vac-method}
\alias{next.ID.state,ID.state.matrix,ID.transition.MSIRV-method}
\title{Methods for Getting State at Next Time Step (next.ID.state methods)}
\usage{
next.ID.state(state, tran, ...)

\S4method{next.ID.state}{ID.state.matrix,ID.transition.SIR}(state, tran)

\S4method{next.ID.state}{ID.state.matrix,ID.transition.SIR.vac}(state, tran)

\S4method{next.ID.state}{ID.state.matrix,ID.transition.MSIRV}(state, tran)
}
\arguments{
\item{state}{current state system (ID.state.matric)}

\item{tran}{ID transition object that has all info for supervising the transition}

\item{...}{ignored}
}
\value{
an updated ID.state.matrix
}
\description{
Methods for Getting State at Next Time Step (next.ID.state methods)
}
