% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{nMx-class}
\alias{nMx-class}
\title{Generate object of class nMx}
\description{
Class definition for transition matrix. This matrix knows how to shift the
ID.state.matrix from one time step to another, and contains all the fun stuff
we need to do the transition (i.e., the matrix)
}
\section{Slots}{

\describe{
\item{\code{rates}}{data.frame.}

\item{\code{mid.age}}{numeric.}
}}

