% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sim.results.MSIRV.R
\name{plot.sim.results.MSIRV}
\alias{plot.sim.results.MSIRV}
\title{Plot method for class "sim.results.MSIRV"}
\usage{
\method{plot}{sim.results.MSIRV}(x, ...)
}
\arguments{
\item{x}{object of class "sim.results.MSIRV"}

\item{...}{ignored}
}
\value{
plot object
}
\description{
Plot method for class "sim.results.MSIRV"
}
