% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sim.results.MSIRV.space.R
\name{plot.sim.results.MSIRV.space}
\alias{plot.sim.results.MSIRV.space}
\title{Plot a sim.results.MSIRV.space object}
\usage{
\method{plot}{sim.results.MSIRV.space}(x, ...)
}
\arguments{
\item{x}{object of class sim.results.SIR}

\item{...}{ignored}
}
\value{
plot of sim results
}
\description{
Plot a sim.results.MSIRV.space object
}
