% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sim.results.SIR.R
\name{plot.sim.results.SIR}
\alias{plot.sim.results.SIR}
\title{Plot function for class sim.results.SIR}
\usage{
\method{plot}{sim.results.SIR}(x, ...)
}
\arguments{
\item{x}{object of class sim.results.SIR}

\item{...}{ignored}
}
\value{
plot of sim results
}
\description{
Plot function for class sim.results.SIR
}
