% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmaternal.R
\docType{methods}
\name{pmaternal}
\alias{pmaternal}
\alias{pmaternal,numeric,maternal.exp.decay-method}
\alias{pmaternal,numeric,maternal.thresh-method}
\title{Methods to get the percent immune by maternal antibodies.}
\usage{
pmaternal(age, mat.obj)

\S4method{pmaternal}{numeric,maternal.exp.decay}(age, mat.obj)

\S4method{pmaternal}{numeric,maternal.thresh}(age, mat.obj)
}
\arguments{
\item{age}{numeric; the age we want the decay rate for}

\item{mat.obj}{the object describing maternally acquired immunity}
}
\value{
percent immune by maternal antibodies
}
\description{
Methods to get the percent immune by maternal antibodies.
}
