% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{prob.vsucc.byage-class}
\alias{prob.vsucc.byage-class}
\title{Object for the probability of vaccine success by age}
\description{
Object for the probability of vaccine success by age
}
\section{Slots}{

\describe{
\item{\code{prob.vsucc}}{numeric.}

\item{\code{ages}}{numeric.}
}}

