% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvacsuccess.R
\docType{methods}
\name{pvacsuccess}
\alias{pvacsuccess}
\alias{pvacsuccess,numeric,vsucc.constant-method}
\alias{pvacsuccess,numeric,vsucc.Logistic-method}
\title{Methods for probability of successful vaccination}
\usage{
pvacsuccess(age, vsucc.obj)

\S4method{pvacsuccess}{numeric,vsucc.constant}(age, vsucc.obj)

\S4method{pvacsuccess}{numeric,vsucc.Logistic}(age, vsucc.obj)
}
\arguments{
\item{age}{the age in months we are checking}

\item{vsucc.obj}{the vaccine success object}
}
\value{
the probability of successful vaccination at that age
}
\description{
Methods for probability of successful vaccination
}
