% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleWAIFW.R
\name{scaleWAIFW}
\alias{scaleWAIFW}
\title{Function to scale the WAIFW matrix to a particular R0 given a particular population}
\usage{
scaleWAIFW(R0, state, waifw, frequency.dep = F, suscept.state = 1)
}
\arguments{
\item{R0}{the reproductive rate to scale to}

\item{state}{a population at disease free equilibrium}

\item{waifw}{the matrix to scale}

\item{frequency.dep}{xxx}

\item{suscept.state}{xxx}
}
\value{
a scaled vertions of waifw
}
\description{
Function to scale the WAIFW matrix to a particular R0 given a particular population
}
