% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleWAIFW.space.R
\name{scaleWAIFW.space}
\alias{scaleWAIFW.space}
\title{Function to scale the WAIFW matrix to a particular R0 given a particular subpopulation, but for a spatially structured population}
\usage{
scaleWAIFW.space(R0, state, waifw, frequency.dep = F, suscept.state = 1)
}
\arguments{
\item{R0}{the reproductive rate to scale to}

\item{state}{a population at disease free equilibrium}

\item{waifw}{the matrix to scale}

\item{frequency.dep}{xxx}

\item{suscept.state}{xxx}
}
\value{
a scaled versions of the age-specific waifw
}
\description{
Function to scale the WAIFW matrix to a particular R0 given a particular subpopulation, but for a spatially structured population
}
