% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{seasonal.age.specific-class}
\alias{seasonal.age.specific-class}
\title{Class for age specific seasonal trend}
\description{
Class for age specific seasonal trend
}
\section{Slots}{

\describe{
\item{\code{amplitude}}{numeric.}

\item{\code{offset}}{numeric.}

\item{\code{period}}{numeric.}

\item{\code{age.class}}{numeric.}

\item{\code{chosen.age}}{numeric.}
}}

