% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{seasonal.age.specific.piecewise.constant-class}
\alias{seasonal.age.specific.piecewise.constant-class}
\title{Seasonal class}
\description{
Designed to take output from TSIR-type analysis
}
\section{Slots}{

\describe{
\item{\code{parameters}}{numeric.}

\item{\code{time.in.year}}{numeric.}

\item{\code{age.class}}{numeric.}

\item{\code{chosen.age}}{numeric.}
}}

