% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{seasonal.cosine-class}
\alias{seasonal.cosine-class}
\title{Class for seasonal cosine}
\description{
Class for seasonal cosine
}
\section{Slots}{

\describe{
\item{\code{amplitude}}{numeric.}

\item{\code{offset}}{numeric.}

\item{\code{period}}{numeric.}
}}

