% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{seasonal.periodic.bspline-class}
\alias{seasonal.periodic.bspline-class}
\title{Class for seasonal periodic bspline}
\description{
Class for seasonal periodic bspline
}
\section{Slots}{

\describe{
\item{\code{parameters}}{numeric.}

\item{\code{nbasis}}{numeric.}

\item{\code{degree}}{numeric.}

\item{\code{period}}{numeric.}
}}

