% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupCountry.Dec2021.R
\name{setupCountry.Dec2021}
\alias{setupCountry.Dec2021}
\title{Function to get all the things you need to run setup coverage and demography for country (NOT Montagu)}
\usage{
setupCountry.Dec2021(country)
}
\arguments{
\item{country}{character; country name}
}
\value{
demography and coverage for each country
}
\description{
Function to get all the things you need to run setup coverage and demography for country (NOT Montagu)
}
