% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupCountry.Nov2023.R
\name{setupCountry.Nov2023}
\alias{setupCountry.Nov2023}
\title{Function to get all the things you need to run setup coverage and demography for country (NOT Montagu)}
\usage{
setupCountry.Nov2023(country)
}
\arguments{
\item{country}{character; country name}
}
\value{
demography and coverage for each country
}
\description{
Function to get all the things you need to run setup coverage and demography for country (NOT Montagu)
}
