% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupCountry.Sep2026.R
\name{setupCountry.Sep2026}
\alias{setupCountry.Sep2026}
\title{Function to get all the things you need to run setup coverage and demography for country (NOT Montagu)}
\usage{
setupCountry.Sep2026(country)
}
\arguments{
\item{country}{character; country name}
}
\value{
demography and coverage for each country
}
\description{
Function to get all the things you need to run setup coverage and demography for country (NOT Montagu)
}
