% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupCountry_202110gavi_v3.R
\name{setupCountry_202110gavi_v3}
\alias{setupCountry_202110gavi_v3}
\title{Function to get all the things you need to run setup coverage and demography for country (Montagu for touchstone 202110gavi_v3)}
\usage{
setupCountry_202110gavi_v3(country)
}
\arguments{
\item{country}{country iso3code}
}
\value{
demography and coverage for each country
}
\description{
Function to get all the things you need to run setup coverage and demography for country (Montagu for touchstone 202110gavi_v3)
}
