% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{sim.results.MSIRV-class}
\alias{sim.results.MSIRV-class}
\title{Holds the results of a simulation with an MSIRV object}
\description{
Holds the results of a simulation with an MSIRV object
}
\section{Slots}{

\describe{
\item{\code{m.inds}}{numeric.}

\item{\code{v.inds}}{numeric.}

\item{\code{routine.intro}}{numeric.}

\item{\code{sia.times}}{numeric.}
}}

