% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{sim.results.MSIRV.space-class}
\alias{sim.results.MSIRV.space-class}
\title{Holds the results of a simulation with a MSIRV.space object}
\description{
Holds the results of a simulation with a MSIRV.space object
}
\section{Slots}{

\describe{
\item{\code{n.subpops}}{numeric.}

\item{\code{subpop.class.label}}{numeric.}
}}

