% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{sim.results.MSIRV.update.demog.space-class}
\alias{sim.results.MSIRV.update.demog.space-class}
\title{Holds the results of a simulation with an experiment.updatedemog object}
\description{
Holds the results of a simulation with an experiment.updatedemog object
}
\section{Slots}{

\describe{
\item{\code{births.each.timestep}}{ANY.}

\item{\code{growth.rate.each.timestep}}{ANY.}
}}

