% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{sim.results.MSIRV.update.demog.vaccine.change-class}
\alias{sim.results.MSIRV.update.demog.vaccine.change-class}
\title{Holds the results output for a simulation with an experiment.updatedemog.vaccinationchange object}
\description{
Holds the results output for a simulation with an experiment.updatedemog.vaccinationchange object
}
\section{Slots}{

\describe{
\item{\code{MR1.fail.each.timestep}}{numeric.}

\item{\code{MR2.fail.each.timestep}}{numeric.}

\item{\code{SIA.fail.each.timestep}}{numeric.}
}}

