% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{sim.results.MSIRV.update.demog.vaccine.change.outbreakresponse-class}
\alias{sim.results.MSIRV.update.demog.vaccine.change.outbreakresponse-class}
\title{Holds the results of an outbreak response simulation}
\description{
Holds the results of an outbreak response simulation
}
\section{Slots}{

\describe{
\item{\code{or.times}}{numeric. Binary vector; 1 at each time step an OBR campaign fired.}
}}

