% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{sim.results.MSIRV.update.demog.vaccine.change.space.school-class}
\alias{sim.results.MSIRV.update.demog.vaccine.change.space.school-class}
\title{Holds the results output for a simulation with an experiment.updatedemog.vaccinationchange.spatial.schoolvaccination object}
\description{
Holds the results output for a simulation with an experiment.updatedemog.vaccinationchange.spatial.schoolvaccination object
}
\section{Slots}{

\describe{
\item{\code{schoolvacc.fail.each.timestep}}{ANY.}
}}

