% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{sim.results.SIR-class}
\alias{sim.results.SIR-class}
\title{Generate of object of class experiment result (sim.result)}
\description{
Class to hold SIR simulation results. Hold the results of the simulation as
SIR compartment values at each time step.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{matrix.}

\item{\code{s.inds}}{numeric.}

\item{\code{i.inds}}{numeric.}

\item{\code{r.inds}}{numeric.}

\item{\code{age.class}}{numeric.}

\item{\code{t}}{numeric.}
}}

