% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRIDTran.R
\name{SIRIDTran}
\alias{SIRIDTran}
\title{Last and final transition function to be called}
\usage{
SIRIDTran(state, tran)
}
\arguments{
\item{state}{xxx}

\item{tran}{xxx}
}
\value{
xxx
}
\description{
Last and final transition function to be called
}
