% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.create.country.x.DFE.ID.state.matrix.R
\name{space.create.country.x.DFE.ID.state.matrix}
\alias{space.create.country.x.DFE.ID.state.matrix}
\title{Function to create a state matrix at the disease free equilibrium (DFE) from a transition object}
\usage{
space.create.country.x.DFE.ID.state.matrix(
  iso3code,
  tot.subpop = NULL,
  tran,
  epi.class.label = c("M", "S", "I", "R", "V"),
  year = 1990
)
}
\arguments{
\item{iso3code}{character; iso3 code}

\item{tot.subpop}{vector; sub-populations by which you can rescale if you want}

\item{tran}{the transition object}

\item{epi.class.label}{epi class label}

\item{year}{numeric; options are 1950-2100}
}
\value{
a starting DFE vector of S and R individuals
}
\description{
Function to create a state matrix at the disease free equilibrium (DFE) from a transition object
}
