% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.create.ID.state.matrix.R
\name{space.create.ID.state.matrix}
\alias{space.create.ID.state.matrix}
\title{Function creates an ID state matrix with the given number of epidemiological categories, age classes, and subpopulations}
\usage{
space.create.ID.state.matrix(
  n.age.class,
  n.epi.class = 3,
  epi.class.label = c("S", "I", "R"),
  n.subpops = 1,
  value = rep(0, n.age.class * n.epi.class * n.subpops)
)
}
\arguments{
\item{n.age.class}{the number of age classes}

\item{n.epi.class}{the number of states in the model}

\item{epi.class.label}{the names of the states}

\item{n.subpops}{xxx}

\item{value}{starting values, defaults to all 0}
}
\value{
an ID.state.matrix object
}
\description{
Function creates an ID state matrix with the given number of epidemiological categories, age classes, and subpopulations
}
