% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.create.ID.transition.MSIRV.R
\name{space.create.ID.transition.MSIRV}
\alias{space.create.ID.transition.MSIRV}
\title{Method to create an ID.transition.MSIRV.space object}
\usage{
space.create.ID.transition.MSIRV(
  n.age.class,
  aging.rate,
  survival.rate,
  waifw,
  routine.vac = 0,
  routine.vac.age.index = 12,
  maternal.obj,
  time.step = 1/2,
  age.class = numeric(),
  birth.rate.subpop = rep(0, n.subpops),
  sia.vac = 0,
  sia.vsucc = new("vsucc.constant", success.rate = 1),
  introduction.rate = 0,
  exponent = exponent,
  frequency.dep = F,
  is.stochastic = F,
  get.births = function(pop, tran) {     return(c(tran@birth.rate, rep(0, length(pop) -
    1))) },
  n.subpops = 1,
  subpop.class.label,
  coupling = 1
)
}
\arguments{
\item{n.age.class}{number of age classes (not location specific)}

\item{aging.rate}{aging rates (includes all locations concatenated)}

\item{survival.rate}{vector with locations survival rates concatenated}

\item{waifw}{assume same across locations}

\item{routine.vac}{xxx}

\item{routine.vac.age.index}{xxx}

\item{maternal.obj}{maternal antibody object}

\item{time.step}{time step in months == generation time}

\item{age.class}{xxx}

\item{birth.rate.subpop}{xxx}

\item{sia.vac}{xxx}

\item{sia.vsucc}{xxx}

\item{introduction.rate}{xxx}

\item{exponent}{xxx}

\item{frequency.dep}{xxx}

\item{is.stochastic}{xxx}

\item{get.births}{xxx}

\item{n.subpops}{number of sub-populations}

\item{subpop.class.label}{xxx}

\item{coupling}{matrix describing movement between sub-populations}
}
\value{
xxx
}
\description{
Method to create an ID.transition.MSIRV.space object
}
