% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Space.Get.CountryX.Starting.Pop.MSIRV.R
\name{Space.Get.CountryX.Starting.Pop.MSIRV}
\alias{Space.Get.CountryX.Starting.Pop.MSIRV}
\title{Function to get starting state vector and transition object - built originally from Get.CountryX.Starting.Pop.MSIRV}
\usage{
Space.Get.CountryX.Starting.Pop.MSIRV(
  uncode,
  generation.time = 0.5,
  age.classes = c(1:240, seq(241, 720, 12)),
  maternal.decay.rt = 0.95,
  exponent = 0.97,
  frequency.dep = TRUE,
  is.stochastic = FALSE,
  tot.subpop = NULL,
  yr.births.per.1000.bysubpop,
  intro.rate = intro.rate,
  flat.WAIFW = FALSE,
  space.asdr.object = NULL,
  targeted.intro = FALSE,
  year = 1980,
  get.births,
  routine.vac = 0,
  routine.vac.age.index = 12,
  n.subpops = 1,
  coupling
)
}
\arguments{
\item{uncode}{UN country code}

\item{generation.time}{the desired generation time in months}

\item{age.classes}{vector; the upper limit of the age classes we are interested in months}

\item{maternal.decay.rt}{rate of maternal decay of immunity to rubella}

\item{exponent}{exponent for the infected}

\item{frequency.dep}{boolean}

\item{is.stochastic}{boolean}

\item{tot.subpop}{numeric vector; population you want to scale you whole experiment by at the beginning by spatial unit, NULL or numeric}

\item{yr.births.per.1000.bysubpop}{numeric vector; crude birth rate per year per 1000 by spatial unit}

\item{intro.rate}{numeric}

\item{flat.WAIFW}{boolean; if FALSE then default waifw is polymod GB}

\item{space.asdr.object}{space.nMx object with rate and mid-age and year and space (age specific death rates)}

\item{targeted.intro}{boolean; FALSE means space introduction out over all age classes, TRUE means to concentrate introductions}

\item{year}{numeric; year to pull country DFE demography (age structure and population size)}

\item{get.births}{vector of births with length of 5*n.age.classes}

\item{routine.vac}{routine vaccination coverage for each subpop}

\item{routine.vac.age.index}{age index based on age classes for each subpop}

\item{n.subpops}{the number of subpopulations}

\item{coupling}{matrix; coupling of subpopulations}
}
\value{
starting state vector and transition object
}
\description{
Function to get starting state vector and transition object - built originally from Get.CountryX.Starting.Pop.MSIRV
}
