% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{space.nMx-class}
\alias{space.nMx-class}
\title{Class definition for space.nMx object - aka Age Specific Death Rates (ASDR) - by space}
\description{
Class definition for space.nMx object - aka Age Specific Death Rates (ASDR) - by space
}
\section{Slots}{

\describe{
\item{\code{rate.years}}{numeric; year associated with each time-specific rate}

\item{\code{rates}}{data.frame; age and space specific death rates per 1 (rows) by year (columns)}

\item{\code{mid.age}}{numeric; mid-age associated with each age-specific rates}

\item{\code{n.subpops}}{numeric; number of subpopulations}
}}

