% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.wrapper.get.routine.time.age.specific.R
\name{space.wrapper.get.routine.time.age.specific}
\alias{space.wrapper.get.routine.time.age.specific}
\title{Wrapper function to add space to get.routine.time.age.specific}
\usage{
space.wrapper.get.routine.time.age.specific(
  time.step = 0.5,
  age.classes = c(1:240, seq(252, 1212, 12)),
  space.time.specific.MR1cov = rbind(rep(0.7, 50), rep(0.6, 50)),
  age.min.MR1 = rep(12, 50),
  age.max.MR1 = rep(23, 50),
  space.time.specific.MR2cov = rbind(rep(0.4, 50), rep(0.3, 50)),
  age.min.MR2 = rep(24, 50),
  age.max.MR2 = rep(35, 50),
  obj.vcdf.MR1 = get.vcdf.uniform(12, 23),
  obj.vcdf.MR2 = get.vcdf.uniform(24, 35),
  obj.prob.vsucc = pvacsuccess(1:(14 * 12), new("vsucc.constant", success.rate = 0.8)),
  MR1MR2correlation = F
)
}
\arguments{
\item{time.step}{numeric, time step in months}

\item{age.classes}{vector, age classes in months, same as tran object}

\item{space.time.specific.MR1cov}{matrix, space (rows) and year (columns) specific routine MR1 coverage as a proportion}

\item{age.min.MR1}{vector, time specific minimum age eligible for MR1}

\item{age.max.MR1}{vector, time specific maximum age eligible for MR1}

\item{space.time.specific.MR2cov}{matrix, space (rows) and year (columns) specific routine MR2 coverage as a proportion}

\item{age.min.MR2}{vector, time specific minimum age eligible for MR2}

\item{age.max.MR2}{vector, time specific maximum age eligible for MR2}

\item{obj.vcdf.MR1}{vaccine.cdf.byage object, age distribution of routine MR1}

\item{obj.vcdf.MR2}{vaccine.cdf.byage object, age distribution of routine MR2}

\item{obj.prob.vsucc}{prob.vsucc.byage object, VE by age relevant to both routine vacccine doses}

\item{MR1MR2correlation}{boolean, if MR1 and MR2 are dependent (TRUE) or independent (FALSE)}
}
\value{
list, features specific to routine vaccination, including age.time.specific.routine matrix which is age and space (rows) by year (cols) of proportion routine vaccination
}
\description{
Wrapper function to add space to get.routine.time.age.specific
}
