% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/space.wrapper.get.schoolvacc.time.age.specific.R
\name{space.wrapper.get.schoolvacc.time.age.specific}
\alias{space.wrapper.get.schoolvacc.time.age.specific}
\title{Wrapper function to add space to get.schoolvacc.time.age.specific}
\usage{
space.wrapper.get.schoolvacc.time.age.specific(
  time.step = 0.5,
  age.classes = c(1:240, seq(252, 1212, 12)),
  space.time.specific.cov = rbind(rep(0.7, 50), rep(0.6, 50)),
  age.min = rep(12, 50),
  age.max = rep(23, 50),
  list.obj.vcdf = replicate(2, list(get.vcdf.uniform(72, 83))),
  obj.prob.vsucc = pvacsuccess(1:(14 * 12), new("vsucc.constant", success.rate = 0.8))
)
}
\arguments{
\item{time.step}{numeric, time step in months}

\item{age.classes}{vector, age classes in months, same as tran object}

\item{space.time.specific.cov}{matrix, space (rows) and year (columns) specific school vaccination coverage as a proportion}

\item{age.min}{vector, time specific minimum age eligible for school vaccination}

\item{age.max}{vector, time specific maximum age eligible for school vaccination}

\item{list.obj.vcdf}{list of vaccine.cdf.byage objects, each item is a subpop, age distribution of school vaccination}

\item{obj.prob.vsucc}{prob.vsucc.byage object, VE by age}
}
\value{
list, features specific to school vaccination, including age.time.specific.school.vaccination matrix which is age and space (rows) by year (cols) of proportion school vaccination
}
\description{
Wrapper function to add space to get.schoolvacc.time.age.specific
}
