% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.wrapper.get.sia.time.age.specific.R
\name{space.wrapper.get.sia.time.age.specific}
\alias{space.wrapper.get.sia.time.age.specific}
\title{Wrapper function to add space to get.sia.time.age.specific}
\usage{
space.wrapper.get.sia.time.age.specific(
  age.classes = c(1:240, seq(252, 1212, 12)),
  space.time.specific.SIAcov = rbind(c(rep(0, 29), 0.5, rep(0, 19), 0.7), c(rep(0, 29),
    0.7, rep(0, 19), 0.8)),
  age.min.sia = rep(12, 50),
  age.max.sia = rep(60, 50),
  obj.prob.vsucc = pvacsuccess(1:240, new("vsucc.constant", success.rate = 0.85))
)
}
\arguments{
\item{age.classes}{vector, age classes in months, same as tran object}

\item{space.time.specific.SIAcov}{matrix, space (rows) and year (columns) specific campaign MR coverage as a proportion}

\item{age.min.sia}{vector, time specific minimum age eligible for MR campaign dose}

\item{age.max.sia}{vector, time specific maximum age eligible for MR campaign dose}

\item{obj.prob.vsucc}{prob.vsucc.byage object, VE by age relevant to campaign vacccine doses}
}
\value{
list, features specific to campaign vaccination, including age.time.specific.SIA matrix which is age and space (rows) by year (cols) of proportion campaign vaccination
}
\description{
Wrapper function to add space to get.sia.time.age.specific
}
