% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.wrapper.surv.prob.over.age.R
\name{space.wrapper.surv.prob.over.age}
\alias{space.wrapper.surv.prob.over.age}
\title{Wrapper function to add space to create.surv.prob.over.age}
\usage{
space.wrapper.surv.prob.over.age(
  age.classes,
  generation.time,
  space.nMx = NULL,
  year = 1980,
  check = FALSE
)
}
\arguments{
\item{age.classes}{vector; a set of age classes for which tran is being built}

\item{generation.time}{numeric; generation time}

\item{space.nMx}{space.nMx object; age and space specific death rates over time}

\item{year}{year to pull age specific death rates that will be used to run out the transients (should coincide with DFE year)}

\item{check}{logical;}
}
\value{
returns the age profile (cols) of survivorship (in units of the generation time) by spatial unit (rows)
}
\description{
Wrapper function to add space to create.surv.prob.over.age
}
