% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.EX.Country.part1.R,
%   R/Spatial.EX.Country.part1.setSUS.R
\name{Spatial.EX.Country.part1}
\alias{Spatial.EX.Country.part1}
\title{Function to set up SPATIAL experiment and run the transients out}
\usage{
Spatial.EX.Country.part1(
  uncode,
  iso3code,
  generation.time = 0.5,
  age.classes = c(1:240, seq(241, 720, 12)),
  maternal.decay.rt = 0.95,
  exponent = 0.97,
  frequency.dep = TRUE,
  tot.subpop = NULL,
  yr.births.per.1000.bysubpop,
  intro.rate = 1e-10,
  targeted.intro = FALSE,
  R0 = 5,
  t.max = 20,
  get.births = NULL,
  seasonal.amp = 0.35,
  flat.WAIFW = FALSE,
  country.specific.WAIFW = FALSE,
  vynnycky.waifw = FALSE,
  vynnycky.waifw.betas = c(2, 1),
  space.asdr.object,
  year = 1990,
  routine.vac = 0,
  routine.vac.age.index = 12,
  n.subpops = 2,
  coupling = matrix(1, nrow = n.subpops, ncol = n.subpops),
  starting.prop.immune = c(0.75, 0.75, 0.75, 0.75),
  starting.prop.immune.ages.in.months = c(9, 12, 24, 240),
  max.immunity = 0.98
)

Spatial.EX.Country.part1(
  uncode,
  iso3code,
  generation.time = 0.5,
  age.classes = c(1:240, seq(241, 720, 12)),
  maternal.decay.rt = 0.95,
  exponent = 0.97,
  frequency.dep = TRUE,
  tot.subpop = NULL,
  yr.births.per.1000.bysubpop,
  intro.rate = 1e-10,
  targeted.intro = FALSE,
  R0 = 5,
  t.max = 20,
  get.births = NULL,
  seasonal.amp = 0.35,
  flat.WAIFW = FALSE,
  country.specific.WAIFW = FALSE,
  vynnycky.waifw = FALSE,
  vynnycky.waifw.betas = c(2, 1),
  space.asdr.object,
  year = 1990,
  routine.vac = 0,
  routine.vac.age.index = 12,
  n.subpops = 2,
  coupling = matrix(1, nrow = n.subpops, ncol = n.subpops),
  starting.prop.immune = c(0.75, 0.75, 0.75, 0.75),
  starting.prop.immune.ages.in.months = c(9, 12, 24, 240),
  max.immunity = 0.98
)
}
\arguments{
\item{uncode}{UN country code}

\item{iso3code}{character iso3 code to matches uncode}

\item{generation.time}{the desired generation time in months}

\item{age.classes}{vector - the upper limit of the age classes we are interested in months}

\item{maternal.decay.rt}{rate of maternal decay of immunity to rubella}

\item{exponent}{exponent for the infected}

\item{frequency.dep}{xxx}

\item{tot.subpop}{population you want to scale you whole experiment by at the beginning, NULL or numeric}

\item{yr.births.per.1000.bysubpop}{crude birth rate per 1000 for each subpopulation}

\item{intro.rate}{numeric - rate at which new infections are introduced}

\item{targeted.intro}{boolean - FALSE means space introduction out over all age classes, TRUE means to concentrate introductions}

\item{R0}{numeric - basic reproductive number of rubella assumed, used to rescale the waifw}

\item{t.max}{numeric - time in year that plan to run out transients}

\item{get.births}{vector of births with length of 5*n.age.classes}

\item{seasonal.amp}{numeric - amplitude of the seasonal cosine function}

\item{flat.WAIFW}{boolean - if false then default waifw is polymod GB}

\item{country.specific.WAIFW}{boolean - if true then overrides polymod or flat waifw}

\item{vynnycky.waifw}{boolean - true if want emilia's waifw <13 and 15-50 (beta_young, beta_old, and 0.7*beta_old)}

\item{vynnycky.waifw.betas}{vector of two betas (beta_young and beta_old)}

\item{space.asdr.object}{nMx object with rate and mid-age (age specific death rates)}

\item{year}{numeric - year to pull DFE demography (population and age structure)}

\item{routine.vac}{xxx}

\item{routine.vac.age.index}{xxx}

\item{n.subpops}{number of subpopulations}

\item{coupling}{matrix of connectivity between subpopulations}
}
\value{
tran object and state.t0 after transients run out

tran object and state.t0 after transients run out
}
\description{
Function to set up SPATIAL experiment and run the transients out

Function to set up SPATIAL experiment and run the transients out
}
