% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial.EX.Country.part2.R
\name{Spatial.EX.Country.part2}
\alias{Spatial.EX.Country.part2}
\title{Function to Run the SPATIAL Simulations/Experiments}
\usage{
Spatial.EX.Country.part2(
  uncode,
  generation.time = 0.5,
  pop.rescale = NULL,
  pop.time = 4,
  is.stochastic = FALSE,
  get.births = NULL,
  t.max = 40,
  rescale.WAIFW = F,
  yr.births.per.1000.acrossyears.bysupop,
  space.asdr.object,
  year = 1990,
  EXt0 = EXt0,
  time.specific.MR1cov,
  time.specific.MR2cov,
  time.specific.SIAcov,
  time.specific.schoolvacc.cov = NULL,
  time.specific.min.age.MR1,
  time.specific.max.age.MR1,
  time.specific.min.age.MR2,
  time.specific.max.age.MR2,
  time.specific.min.age.SIA,
  time.specific.max.age.SIA,
  time.specific.min.age.schoolvacc,
  time.specific.max.age.schoolvacc,
  obj.vcdf.MR1 = get.vcdf.normal(6, 12),
  obj.vcdf.MR2 = get.vcdf.normal(15, 21),
  list.obj.vcdf.schoolenroll,
  obj.prob.vsucc = pvacsuccess(1:(20 * 12), get.boulianne.vsucc()),
  sia.timing.in.year = (3/12),
  schoolvacc.timing.in.year = (9/12),
  MR1MR2correlation = NULL,
  MR1SIAcorrelation = NULL,
  MR2SIAcorrelation = NULL,
  SIAinacc = NULL,
  prop.inacc = NULL,
  SIAinefficient = NULL
)
}
\arguments{
\item{uncode}{UN country code}

\item{generation.time}{generation time in months}

\item{pop.rescale}{numeric - population by which you want to rescale at pop.time}

\item{pop.time}{numeric - time in YEARS that you want to rescale the population by}

\item{is.stochastic}{logical}

\item{get.births}{vector of births with length of 5*n.age.classes}

\item{t.max}{numeric - time in year that plan to run the experiment}

\item{rescale.WAIFW}{logical}

\item{yr.births.per.1000.acrossyears.bysupop}{vector of length(t.max) - crude birth rate per 1000 per year}

\item{space.asdr.object}{space nMx object - with rate and mid-age (age specific death rates)}

\item{year}{numeric - start year}

\item{EXt0}{object from part1}

\item{time.specific.MR1cov}{- numeric matrix}

\item{time.specific.MR2cov}{- numeric matrix}

\item{time.specific.SIAcov}{- numeric matrix}

\item{time.specific.schoolvacc.cov}{- numeric matrix}

\item{time.specific.min.age.MR1}{- numeric vector}

\item{time.specific.max.age.MR1}{- numeric vector}

\item{time.specific.min.age.MR2}{- numeric vector}

\item{time.specific.max.age.MR2}{- numeric vector}

\item{time.specific.min.age.SIA}{- numeric vector}

\item{time.specific.max.age.SIA}{- numeric vector}

\item{time.specific.min.age.schoolvacc}{- numeric vector}

\item{time.specific.max.age.schoolvacc}{- numeric vector}

\item{obj.vcdf.MR1}{- object}

\item{obj.vcdf.MR2}{- object}

\item{list.obj.vcdf.schoolenroll}{- list of objects}

\item{obj.prob.vsucc}{- object}

\item{sia.timing.in.year}{- numeric scalar}

\item{schoolvacc.timing.in.year}{- numeric scalar}

\item{MR1MR2correlation}{- NULL or not}

\item{MR1SIAcorrelation}{- NULL or not}

\item{MR2SIAcorrelation}{- NULL or not}

\item{SIAinacc}{- xxx}

\item{prop.inacc}{- xxx}

\item{SIAinefficient}{- xxx}
}
\value{
experiment result
}
\description{
Function to Run the SPATIAL Simulations/Experiments
}
