% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toProportions.R
\docType{methods}
\name{toProportions}
\alias{toProportions}
\alias{toProportions,sim.results.MSIRV-method}
\title{Convert the results to proportions of each age class in each compartment}
\usage{
toProportions(ob, ...)

\S4method{toProportions}{sim.results.MSIRV}(ob)
}
\arguments{
\item{ob}{object of type sim.results.MSIRV}

\item{...}{ignored}
}
\value{
result object of proportions instead of absolute numbers
}
\description{
Convert the results to proportions of each age class in each compartment
}
