% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{vacc.per.time.step.by.age-class}
\alias{vacc.per.time.step.by.age-class}
\title{Generate object of class vacc.per.time.step.by.age}
\description{
Class that holds the information on how many people should be
vaccinated during each time step in each age class
in the TSIR framework.
}
\section{Slots}{

\describe{
\item{\code{pvacc.in.age.class}}{numeric.}
}}

