% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{vaccine.cdf.byage-class}
\alias{vaccine.cdf.byage-class}
\title{Generate and object of class vaccine.cdf.byage}
\description{
Object for the CDF of vaccination by age (non scaled)
}
\section{Slots}{

\describe{
\item{\code{cdf}}{numeric.}

\item{\code{ages}}{numeric.}
}}

