% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{vsucc.constant-class}
\alias{vsucc.constant-class}
\title{Make a constants vaccine success object}
\description{
Make a constants vaccine success object
}
\section{Slots}{

\describe{
\item{\code{success.rate}}{numeric.}
}}

