% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClasses.R
\docType{class}
\name{vsucc.Logistic-class}
\alias{vsucc.Logistic-class}
\title{Make an object to hold a logistic function for vaccine success}
\description{
Make an object to hold a logistic function for vaccine success
}
\section{Slots}{

\describe{
\item{\code{intercept}}{numeric}

\item{\code{mo.eff}}{numeric}

\item{\code{full.efficacy}}{numeric}
}}

