% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertAgeSIA.R
\name{convertAgeSIA}
\alias{convertAgeSIA}
\title{Function to convert the SIA age range into something usable}
\usage{
convertAgeSIA(sia.age.range)
}
\arguments{
\item{sia.age.range}{input as AgeGroup from the WHO SIA Data}
}
\value{
Returns age range
}
\description{
Function to convert the SIA age range into something usable
}
