% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.routine.time.age.specific.R
\name{get.routine.time.age.specific}
\alias{get.routine.time.age.specific}
\title{Get Routine time age-specific}
\usage{
get.routine.time.age.specific(
  time.step = 0.5,
  age.classes = c(1:240, seq(252, 1212, 12)),
  time.specific.MR1cov = rep(0.6, 50),
  age.min.MR1 = rep(12, 50),
  age.max.MR1 = rep(23, 50),
  time.specific.MR2cov = rep(0.5, 50),
  age.min.MR2 = rep(24, 50),
  age.max.MR2 = rep(35, 50),
  obj.vcdf.MR1 = get.vcdf.uniform(12, 23),
  obj.vcdf.MR2 = get.vcdf.uniform(24, 35),
  obj.prob.vsucc = pvacsuccess(1:(14 * 12), new("vsucc.constant", success.rate = 0.8)),
  MR1MR2correlation = F
)
}
\arguments{
\item{time.step}{xxx}

\item{age.classes}{xxx}

\item{time.specific.MR1cov}{xxx}

\item{age.min.MR1}{xxx}

\item{age.max.MR1}{xxx}

\item{time.specific.MR2cov}{xxx}

\item{age.min.MR2}{xxx}

\item{age.max.MR2}{xxx}

\item{obj.vcdf.MR1}{xxx}

\item{obj.vcdf.MR2}{xxx}

\item{obj.prob.vsucc}{xxx}

\item{MR1MR2correlation}{xxx}
}
\value{
list for routine time
}
\description{
Get Routine time age-specific
}
