% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.schoolvacc.time.age.specific.R
\name{get.schoolvacc.time.age.specific}
\alias{get.schoolvacc.time.age.specific}
\title{Get School Vaccination time age-specific}
\usage{
get.schoolvacc.time.age.specific(
  time.step = 0.5,
  age.classes = c(1:240, seq(252, 1212, 12)),
  time.specific.cov = rep(0.6, 50),
  age.min = rep(12, 50),
  age.max = rep(23, 50),
  obj.vcdf = get.vcdf.uniform(12, 23),
  obj.prob.vsucc = pvacsuccess(1:(14 * 12), new("vsucc.constant", success.rate = 0.8))
)
}
\arguments{
\item{time.step}{xxx}

\item{age.classes}{xxx}

\item{time.specific.cov}{xxx}

\item{age.min}{xxx}

\item{age.max}{xxx}

\item{obj.vcdf}{xxx}

\item{obj.prob.vsucc}{xxx}
}
\value{
list for routine time
}
\description{
Get School Vaccination time age-specific
}
