% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNumber.per.AgeYear.R
\name{GetNumber.per.AgeYear}
\alias{GetNumber.per.AgeYear}
\title{Function to get number of individuals per age in years}
\usage{
GetNumber.per.AgeYear(vec, age.classes)
}
\arguments{
\item{vec}{vector of length age.classes}

\item{age.classes}{age.classes}
}
\value{
number of individuals per age in years
}
\description{
Function to get number of individuals per age in years
}
