% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\docType{methods}
\name{run}
\alias{run}
\alias{run,experiment.updatedemog-method}
\alias{run,experiment.updatedmog,ANY-method}
\alias{run,experiment.updatedemog.vaccinationchange-method}
\alias{run,experiment.updatedmog.vaccinationchange,ANY-method}
\alias{run,experiment.updatedemog.vaccinationchange.vaccinationlimitations-method}
\alias{run,experiment.updatedemog.vaccinationchange.vaccinationlimitations,ANY-method}
\alias{run,experiment.updatedemog.vaccinationchange.spatial-method}
\alias{run,experiment.updatedemog.vaccinationchange.spatial,ANY-method}
\alias{run,experiment.updatedemog.spatial-method}
\alias{run,experiment.updatedemog.spatial,ANY-method}
\alias{run,experiment.updatedemog.vaccinationchange.vaccinationlimitations.spatial-method}
\alias{run,experiment.updatedemog.vaccinationchange.vaccinationlimitations.spatial,ANY-method}
\title{Methods to run experiment}
\usage{
run(exper, ...)

\S4method{run}{experiment.updatedemog}(exper, rescale.WAIFW = TRUE, ...)

\S4method{run}{experiment.updatedemog.vaccinationchange}(exper, rescale.WAIFW = TRUE, ...)

\S4method{run}{experiment.updatedemog.vaccinationchange.vaccinationlimitations}(exper, rescale.WAIFW = TRUE, ...)

\S4method{run}{experiment.updatedemog.vaccinationchange.spatial}(exper, rescale.WAIFW = T, ...)

\S4method{run}{experiment.updatedemog.spatial}(exper, rescale.WAIFW = T, ...)

\S4method{run}{experiment.updatedemog.vaccinationchange.vaccinationlimitations.spatial}(exper, rescale.WAIFW = T, ...)
}
\arguments{
\item{exper}{the experiment object}

\item{...}{additional arguments}

\item{rescale.WAIFW}{logical; should we rescale WAIFW matrix}
}
\value{
an experiment result object
}
\description{
Methods to run experiment
}
