% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.getDemography.R
\name{space.getDemography}
\alias{space.getDemography}
\title{Function to get spatial demography - this function is temporary and will need to be totally revamped}
\usage{
space.getDemography(uncode)
}
\arguments{
\item{uncode}{UN country code}
}
\value{
list, demographic data
}
\description{
Function to get spatial demography - this function is temporary and will need to be totally revamped
}
