% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.getDemography.R
\name{space.getDemography}
\alias{space.getDemography}
\title{Function to get spatial demography}
\usage{
space.getDemography(age.classes, iso3code)
}
\arguments{
\item{age.classes}{numeric - in months}

\item{iso3code}{character}
}
\value{
list, demographic data
}
\description{
Function to get spatial demography
}
